+++ morpheus/messages/message_meta.py
@@ -1,0 +1,48 @@
+# Copyright (c) 2021-2022, NVIDIA CORPORATION.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import dataclasses
+
+import pandas as pd
+
+import morpheus._lib.messages as neom
+from morpheus.messages.message_base import MessageBase
+
+
+@dataclasses.dataclass
+class MessageMeta(MessageBase, cpp_class=neom.MessageMeta):
+    """
+    This is a container class to hold batch deserialized messages metadata.
+
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        Input rows in dataframe.
+
+    """
+    df: pd.DataFrame
+
+    @property
+    def count(self) -> int:
+        """
+        Returns the number of messages in the batch.
+
+        Returns
+        -------
+        int
+            number of messages in the MessageMeta.df.
+
+        """
+
+        return len(self.df)
